Make intelligent patching mode 4 controllable with the ctrl modifier.
Intelligent patching mode 4 often just always prefers option A (outgoing
connections). But if the user actually wants option B (incoming connections),
he can't get it because option A was chosen. This is confusing, which is why
mode 4, while very useful if it works as intended, can be rather difficult to
use right now.

We therefore now take into account the status of the ctrl modifier, which
simply reverses our choice of option A or B, so that the user now has a way of
getting option B even if option A is preferred (and vice versa). If the ctrl
modifier isn't pressed, mode 4 works exactly as before, so this change is
fully backwards-compatible.