Overhaul the license files for more extensive and clearer license information.
Clarify the license terms of the package as well as its bundled components,
and place the original license files into the main directory of the source, so
that everybody can find them easily. The source is in Markdown format for
easier editing (LICENSE.md), from which the html file (LICENSE.html) can be
generated with: pandoc LICENSE.md -o LICENSE.html. The html file is now used
in packages/Makefile to generate the License.html file to be included in the
package.