remove copies of mapping library abstractions from la-kitchen
I don't see any use of the namespace prefix in the repo (like
la-kitchen/amplitude_n). So while this has the potential to break
some patches in the wild, it appears most people have been using
either unprefixed names or the more popular mapping library prefix.

If this ends up breaking a bunch of stuff we can always revert it.