use int instead of t_int
It appears int and t_int were freely mixed in this external, causing a size
discrepancy that can result in a buffer overflow. Since t_int is supposed to
be function pointer-sized container, it's unclear why that would be used here.
This commit changes all t_int use to int.

If t_int was really meant here for some reason then we can go in the other
direction.