hook in the functionality to suppress scrolling.
However, we still need to figure out if we want to supress gui_canvas_getscroll
when suppressing the scrollbars. Currently it still gets called, which will
adjust the viewport to make objects at negative indices visible. I think
instead we want to anchor the viewport to 0,0 at top left and supress those
calls to get a nice speedup in the GUI.