speed up msg box by caching the glist in t_messresponder for fetching $0
This brings msg box back to the same performance it had before $0 expansion
was added as a feature to message boxes.

The first implementation relied on calling canvas_setcurrent and
canvas_unsetcurrent in the message box methods. This added a malloc/free
for each message box method call.

Instead this implementation caches the msg box glist in the t_messresponder.
That is available as the target parameter in m_binbuf. Unfortunately, we
have to add a parameter to binbuf_realizedollsym. Since that's a public
function we have to break out the new param into binbuf_dorealizedollsym.