add regression tests for changes to message box (and future regressions)
This is an initial attempt at a regression testing framework.

Initially it has some regression tests that come from the work on
improvements to the msg box.

The framework allows the tests themselves to remain extremely simple. They
just need to send a number and potential error message to the outlet.

However, they do need to be added in the main patch. This is because we
will need to run some tests in a certain order. We could number the test
filenames but that's less easy to change later.