use a macro to set GUIDIR
This makes it easier to point the gui dir at a local copy of the repo for
easier development. You can use this to make it easier to test out changes
to the GUI in purr-data/pd/nw.

Setting GUIDIR should even work if you haven't installed Purr Data to the
system. Just make sure the nw.js binary has been downloaded and expanded to
purr-data/pd/nw/nw. (This happens automatically when you build.)