Invoke desktop actions via /bin/sh.
We do this to prevent extra arguments to be passed from the DE.

At least on KDE, it seems that the desktop file itself gets passed if the
action is invoked by right-clicking the icon on the desktop (this doesn't
happen with panel icons). Maybe this is a bug in KDE, but we need to prevent
this, because otherwise Purr Data tries to open the desktop file as a patch,
which of course will cause a bunch of ugly error messages in the main window.