First checkpoint in implementing custom scrollbars and improving the scrollbar behavior. This is a WIP that requires nw.js 0.46.2.
Finished second stage of the scrollbar implementation and added scrolling with middle click. All scrollbars are now operational and to the best of my knowledge accurate. The only things remaining are: getscroll needs to be called on deselect to recalculate when the handles are erased on objects that have them (we should get rid of these and focus on bboxes instead), responding to maximizing and restoring, and addressing a weird bug where in 0.46.2 patches are saved with larger window sizes under certain conditions.

Reverted scrolling by grabbing due to problematic relationship between scrollBy and the zoom factor

Continuing work on the scrollbars. The only known thing now remaining is ensuring that the patch window is properly recorded when saving it (under certain conditions it spits out completely wrong values

Removed debug printouts

Finished the scrollbars. The only thing remaining is for someone who has a better understanding of the pd_canvas.js and how the events are captured, to refactor my registration of events that ensure scrollbars are clickable and the onresize event. Currently, they are embedded inside the pd_canvas.js under the body tag where you will find both onscroll and onresize events.

Overhaul of the new scrollbars to further improve their behavior, enable as perfect of a fit of the plots both in gop and on toplevel, and dynamic updating of the plots in toplevel when resized. Reworked dynamic event adding/removal for the scrollbars and removed onscroll from the html document and moved it into the pd_canvas.js. Still need to add a delay for the backend callback to update the scrollbars on the toplevel plot/array and thereby remove scrollbar flicker.

Finished implementing the scrollbar callback and addressed popup not appearing in the right place (likely due to upgrade to 0.46.2)

Added names of arrays to their toplevel subpatch window title. Multiple arrays are listed in the order they have been added.

Another reworking of scrollbars to and array sizing to capture all possible scenarios. Some refactoring to make use of the canvas_hasarray. Remaining known issues pertaining to the plots are:
*array toplevel windows are not restored in position they were saved
*last element on the plots is not clickable
*make bezier plot look like bezier (may need to leverage curve_path)