fix regression caused by unnecessarily unvis'ing gobjs in graph_vis
This was initially part of an effort to prevent double binding to scalars
for certain event callbacks. By unvis'ing all scalars in the graph this
would ensure the symbol would get unbound.

However, we're already doing that in canvas_map. So it appears this extra
unvis'ing in graph_vis was unnecessary.

However, it looks like the scalar in a gop will get unvis'd twice when
destroying the gop canvas. No idea why that is, but if there's a regression
perhaps one of the few people using scalar events will report a regression :)