pd vanilla fwd intlet~ argument port
comment the line that defines an extra inlet for the 'inlet~' object

show error message in the case its a inlet(~) object but argument 'fwd' is not specified. Third approach

fix for case fwd message is not properly built, pd could crash

add zcheckgetfn method. given a selector and a list of param types, checks if the object has that method and if the params coincide

switch to mess3 method for calling fwd callback and add proper argument verification,  pack conditions to call inlet_fwd inside it, selector 'fwd' can be sent again through inlet objects

create wrappers of symbol and blob for inlet_fwd

create wrappers of pointer for inlet_fwd