add hex symbol color syntax for iemgui colors
We're not yet saving the hex colors as that will break in older patches.
But we do print out warning messages in this interim period and will
save the hex colors at a later date. Once that change is made, all the
interim versions will be able to read the new format