Fix some remaining issues with the touch interface.
- Disable touch-scrolling in situations where it interferes with other
  actions, like operating a slider, or dragging objects or the selection
  rectangle in edit mode.

- Truncate fractional touch coordinates which were causing glitches in the
  FUDI interface.

- Make dragging of iemgui labels and scalar draggables work (kind of).

TODO: Dragging (specifically in edit mode) can be a bit tricky in cases where
it's difficult to hit the right spot to initiate the drag action. This
*might* be solvable by making the algorithm identifying these spots a bit more
lenient, and accept touches which are a bit off.

Also, despite my best efforts, dragging iemgui label handles in edit mode
still needs quite some faffing about when using touch: After clicking the
iemgui object to make the handle appear, you first need to drag the handle a
little (horizontally, vertically, or diagonally) to make the corresponding
drag cursor appear, then release and touch the drag cursor *again* to drag the
handle to its final destination (and the label along with it). It goes without
saying that this seems overly complicated, but there doesn't seem to be an
easy way to work around this in the GUI, as the whole process goes back and
forth between GUI and engine in several steps in the current implementation.