List devices from -alsaadd before hardware devices
If a user added a device on the commandline, it's likely that they want
this device to be the default.

This allows Purr Data to work nicely inside a Flatpak sandbox, as we
can use `-alsaadd pulse` to enable audio output to PulseAudio by
default, with direct hardware access still an option.