Add user-defined GUI presets.
This discovers user-defined GUI presets (css files) in the css
subdirectory of the gui folder, and adds them at the end of the
corresponding dropdown menu in the GUI prefs. It is a convenience
intended for users and developers who want to experiment with their own
GUI styles.

The predefined presets are still hardcoded in the dialog_prefs.html
file, and are always available (and displayed first in the dropdown
menu). If the preset saved in the prefs doesn't exist anymore (e.g.,
because the user removed its css file), it will be ignored at startup
and the default style will be used instead.

Note that this isn't entirely foolproof since the fallback may not exist
if the user also removed the default.css file. But this could also
happen previously when the css directory gets messed up, and in such a
case the user simply gets what he or she deserves.