Add a clickable bookmark indicator.
This takes the form of a little bookmark icon to the right of the file
browser icon. Just the plain icon indicates that the current directory
hasn't been bookmarked yet, and that clicking on it will bookmark it.
If you do this, a little red cross appears on the icon, which reminds
you that the directory has been bookmarked, and that clicking on the
icon will remove the bookmark again.

Note that in general a directory may have been bookmarked multiple
times, by using the keyboard shortcut (Ctrl-D). In this case the red
cross will stick until the last instance of the bookmark has been
removed.

The icons come from the Manjaro KDE Breath theme (a variation of the
KDE Breeze theme), and we also replaced the folder icon with a matching
icon from the same theme. Thanks are due to the Manjaro and KDE teams!

While we were at it, we also fixed the vertical alignment of the icons
with respect to the search field, they are properly aligned to the
middle now.