Fix case-sensitive search in the browser's findbox.
This uses findAndReplaceDOMText(), which apparently looks for
case-sensitive matches, but the search text entered by the user is
*always* converted to lowercase. This means that mixed case search terms
just won't be found, which is bad. (E.g., try searching for "GEM" in the
toc to see this bug in action.)

Fixed by just removing the call to toLowerCase() on the search text,
case-sensitive searches now work as expected, problem solved.