Add a global method to keep the watchdog happy while the GUI is busy.
For some reason, GUI operations which run for extended periods of time
may cause the engine to crash with a SIGHUP error on Linux. Apparently
this is caused by the watchdog bailing out and taking the engine with
it. This affects the new print operation as well as the existing message
operation in the file menu, possibly also search index generation for
the help browser, and maybe some other operations that still need to be
identified.

The new gui-busy method lets the GUI set an internal flag in the engine
which makes the engine keep the watchdog happy while the GUI does its
thing, using the same mechanism that is already in place to ping the
watchdog when running GUI-less. To these ends, the GUI needs to invoke
"pd gui-busy 1" when starting the time-consuming operation, and "pd
gui-busy 0" when the operation is done.