comment out libdl.dll requirement as it no longer ships with msys2
Not sure what this was used for. Could have been a remnant from ancient
Pd-extended builds where libdl was required to dynamically load at all,
but that's a shot in the dark.

Anyway, if we get reports in a future release of some external randomly
refusing to load under Windows, we may want to test whether this change
is the culprit.