mingw64: Remove libdl.dll from the package.
NOTE: We've already done this for the mingw32 package in rev. 807068ce, which
was prompted by a recent upgrade of the dlfcn package in msys2 which no longer
ships that dll.

However, it seems that the new dlfcn package also breaks Gem. At present,
no-one knows what the actual issue with Gem is. But for the time being we have
to downgrade the mingw-w64-i686-dlfcn package from 1.2.0-2 to 1.2.0-1 (which
is known to work with Gem), until a future update of msys2 fixes the issue,
or we figure out a way to work around it in Gem.