Add a canvas-local zoom flag.
This adds a canvas member variable and the necessary logic so that
loading and saving zoom can be enabled and disabled in a canvas-local
fashion via `declare -zoom`, overriding the global sys_zoom flag.

An update of the declare help patch is included. It adds a new
Patch_local subpatch which explains the new option.