pddplink: handle symbol input on the 1st inlet.
As discussed on the mailing list, it's useful to allow symbol
inputs (other than bang) on pddplink's inlet, to make it easier to
invoke pddplink on a variable symbol (filename, patch, URL) value.

Note that for backward compatibility, a dummy filename argument is still
needed. Please check the updated help patch for details and an example.