Add related objects to the help index
If an object has a "related objects" field in its help patch meta data,
then the list of related objects will be added to the help index and
will be displayed along with its description in the help browser.

Moreover, objects will be listed in the search results if they have the
search term as a related object. For instance, a search for 'pulse' will
also list 'metro' because 'metro' has 'pulse' as a related object.

Note that depending on the actual meta data, relatedness isn't
necessarily symmetric. E.g., 'pulse' doesn't list 'metro' as a related
object in turn and thus a search for 'metro' will *not* show 'pulse' in
the search results, even though 'pulse' is related to 'metro'.