Add link to related_objects to each -help file
This adds clickable links (employing calls to pdgui.doc_open) to the
help patches for related objects, if we can locate them. To support
this, we add an initial scan of the help patches, recording their names
and paths in the search index so that we can quickly look up their
locations.

NOTE: At present, the initial scan only covers the default hierarchy in
libdir, regardless of what the actual scope of the browser set in the
gui preferences and the help path is. Meta data will only be retrieved
for the help patches discovered in the initial scan.