Add a keyboard command to remove completions from the index.
There are reasons why you may want to get rid of completions. Sometimes
the autocompletion index may contain completions that are not actually
valid object names, or arguments that you may no longer wish to use.
This commit introduces a new key binding for ctrl+y which lets you
remove a completion from the index. (Note that ctrl+y is also used for
the Tidy Up operation in the Edit menu, but the two bindings don't
interfere with each other.)

If a valid completion has been entered, ctrl+y will ask in the Pd
console to confirm that you really want to remove the completion.
Pressing ctrl+y again will remove the completion and report this in the
console. Pressing Esc or any other key will cancel the operation.

Note that we keep this low-profile by design. While a GUI implementation
would be possible, the keyboard + console interface is quick and
unobtrusive, and also much easier to implement.