Make sure that the index is rebuilt if autocompletion needs it.
As generation of the autocompletion index is tied in with the generation
of the help index, building the index can't be deferred until the help
browser is opened if autocompletion is enabled. Previously this resulted
in missing completions or autocompletion being completely unavailable
until the help index was (re-)built. This commit addresses this bug by
ensuring that the help index is generated right away when needed.