Sort completion results.
Fuse's result lists are ordered according to similarity scores, which
may look pretty arbitrary at times. Since we only display at most 8
results at a time, this is a problem, because a rather random selection
of completions may be shown in the menu. To improve this, we now also
order results for the same score alphabetically, which makes the results
look much more sensible IMHO.