Make the auto-complete popup follow along with the current Tab completion.
This effectively gives a way to scroll the popup using Tab to cycle
through the completions, without having to add a scrollbar to the menu.

While we were at it, we also added some new keybindings for Alt+Home and
Alt+End which quickly take you to the first and last completion,
respectively. It's easy to get lost in the completion list if it gets
very long (try completing `list` some time), so this makes navigating
the list much easier.