Add an option to sort completions by relevance.
If enabled, this orders object and argument completions first by
relevance, then alphabetically by object name, with "relevance"
currently being defined as frequency of use (as given by the occurrence
field in the completion data).

As usual, the option is stored in the user preferences, its checkbox can
be found on the GUI Preferences tab.