Improved sort critera for relevance (hopefully).
Built-in flag and live usage data are now combined into a single numeric
relevance score, with a boost factor determining the weight of the
former relative to the latter. This makes it easy to adjust the
relevance order. Currently we have the boost factor at 50, which means
that an external needs to have 51 uses before it gets preferred over a
built-in with no uses at all. Is this too large? Too small? Also, should
the fuse similarity score take precedence over relevance (currently it
doesn't)? At present nobody knows, so let's give it a try and listen to
user feedback to improve it.