Get rid of the python2 build dependency.
Python 2 has been EOL since 2019, and isn't available in the Arch
repositories any more, so continuing to support it has become a major
hassle. The quick and dirty fix is to just build the DISIS cwiid library
without it, which seems to be the only part of the source which still
depends on Python 2 at build time.