M1 fixes - fix up external Makefiles.
Specifically, we disable the branches which attempt to detect iOS
cross compilation by checking for a Darwin system with arm cpu
architecture. With Apple Silicon Macs this test is now broken, since
an M1/M2 Mac would be detected as an iOS target.

For now, we skip this check by looking for a non-existent arm-iphone
architecture. This should be fine since currently we don't support iOS
anyway, and AFAIK Purr Data has never been tested there. If we want to
revive iOS cross compilation at some point, we'll have to devise a
check which will do the right thing on both iOS and macOS arm devices.