M1 fixes - rework Homebrew detection.
On Apple Silicon, Homebrew is installed in a different location. Thus
in the externals Makefile and the embed-MacOSX-dependencies.sh script,
we now check the HOMEBREW_PREFIX environment variable to determine
whether Homebrew is installed and, if so, where it is.

Also added missing include and library paths in some externals so that
they actually find the Homebrew header files and libraries.