Keep track of the actual device names in the MIDI prefs.
This adds new midiindevname and midioutdevname keys to the preferences,
which enables us to remap device indices on the fly at startup in case
the device list has changed during invocations. (Which happens a lot, in
particular, with the portmidi backend, due to MIDI devices being plugged
and unplugged, change of virtual MIDI devices, etc.)

Note that this doesn't provide hotplugging support, which is really what
we'd like to have, but is currently impossible due to backend
limitations. But at least Purr Data will not connect to the wrong
devices on relaunch any more, which previously was a major annoyance.