Overhaul of the hammereditor implementation.
This was completely rewritten in order to minimize the changes to the
existing text dialog interface. In particular, the text buffer used in
hammereditor_append is now maintained on the C side. This additional
buffer is required by the cyclone objects which call this routine to
populate the hammereditor before the text dialog window is mapped, so
gui_text_dialog_append doesn't work in that context.

The only essential change in the text dialog interface is in the way the
window position gets passed to the gui_text_dialog routine. We have two
new x, y arguments there which are necessary to make the click actions
of various cyclone objects work. The text object in x_text.c, which is
the only other object where this gui call is being used, was modified
accordingly, and seems to work alright with this change.